Create README.md
I leveraged on Tableau software to provide a comprehensive overview of the African Mobile's performance across key metrics. This interactive dashboard presents crucial insights on Profit by City, Profit by Segment, and Profit Trend over the past four years of the company's operations. 

The map on the dashboard displays Profit by City, allowing stakeholders to visually assess the contribution of each city to the overall profitability. Each city is represented by a marker, with the size of the marker corresponding to the magnitude of profit. Hovering over a city marker reveals detailed information, including the exact profit amount. 

The bar chart breaks down the profitability by segment, providing a clear comparison between different product or service segments. Stakeholders can easily identify which segments are driving the highest profits and make informed decisions based on these insights. The line chart illustrates the Profit Trend over the past four years, offering a dynamic view of the company's financial performance. The ability to observe trends enables stakeholders to identify patterns and make strategic decisions to capitalize on growth opportunities or address challenges.

One of the key features of this dashboard is its interactivity. By selecting a specific city on the map, stakeholders can see a real-time adjustment of the other charts, providing a localized perspective on Profit by Segment and the Profit Trend for the chosen city. This feature facilitates a more granular analysis tailored to the unique characteristics of each market.